Extracted persistance interface from domain
