+++ platform/persistance/src/main/java/io/dotanuki/norris/persistance/SearchHistoryInfrastructure.kt
@@ -1,13 +1,12 @@
 package io.dotanuki.norris.persistance
 
 import android.content.SharedPreferences
-import io.dotanuki.norris.domain.services.SearchesHistoryService
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
 internal class SearchHistoryInfrastructure(
     private val prefs: SharedPreferences
-) : SearchesHistoryService {
+) : LocalStorage {
 
     override suspend fun lastSearches(): List<String> {
         return suspendCoroutine { continuation ->
